feat(migration): 新增 templates 資料表及預設模板資料
